Renamed namespace IdentityDTO into AuthDTO
